requirements.txt
streamlit
scikit-learn
pandas
numpy
python-docx
docx2txt
PyMuPDF
